+++ src/miejsca.cpp
@@ -8,7 +8,7 @@ Miejsce::Miejsce()
     cout << "Poprawnie stworzono konstruktor (Miejsca)!\n";
 }
 
-void Miejsce::set_miejsce(string n, int odl)
+void Miejsce::set_miejsce(string n, double odl)
 {
     this->nazwa = n;
     this->odleglosc = odl;
@@ -19,7 +19,7 @@ string Miejsce::get_nazwa(void)
     return this->nazwa;
 }
 
-int Miejsce::get_odleglosc(void)
+double Miejsce::get_odleglosc(void)
 {
     return this->odleglosc;
 }
