+++ src/pociagi.cpp
@@ -8,7 +8,7 @@ Pociag::Pociag()
     cout << "Poprawnie stworzono konstruktor (Pociagi)!" << endl;
 }
 
-void Pociag::set_pociag(string n, int p)
+void Pociag::set_pociag(string n, double p)
 {
     this->nazwa = n;
     this->predkosc = p;
@@ -19,7 +19,7 @@ string Pociag::get_nazwa(void)
     return this->nazwa;
 }
 
-int Pociag::get_predkosc(void)
+double Pociag::get_predkosc(void)
 {
     return this->predkosc;
 }
